refactor(tui): consolidate duplicated hover, truncation, and settings helpers
Fold four copies of the same logic into shared homes, prompted by the
settings-search work touching all of them:

- settings/render.rs carried a private paint_hover_bg identical to the
  one every dialog already uses from components::hover; use the shared
  one.
- width-aware truncate-with-ellipsis existed privately in home/render.rs
  and was re-implemented inline in the search popup's value rows; both
  now use components::truncate_to_width (new components/text.rs, tests
  moved along, plus a wide-glyph case).
- the (scope, base, overrides) triple for build_fields_for_category was
  matched out twice in settings/mod.rs; extracted field_build_inputs().
- the popup hit-row lookup was duplicated between click and hover
  routing; extracted search_hit_at_row().
- four identical fresh_view test fixtures across the settings input and
  render test modules collapsed into settings::test_util::fresh_view.

No behavior change.
