+++ src/tui/components/mod.rs
@@ -9,6 +9,7 @@ pub(crate) mod hover;
 mod list_picker;
 pub(crate) mod preview;
 pub(crate) mod scroll;
+mod text;
 mod text_input;
 mod tool_config;
 
@@ -17,6 +18,7 @@ pub use dir_picker::{DirPicker, DirPickerResult};
 pub use help::HelpOverlay;
 pub use list_picker::{ListPicker, ListPickerResult};
 pub use preview::{format_scroll_indicator, Preview};
+pub use text::truncate_to_width;
 pub(crate) use text_input::{focused_input_spans, input_scroll, visible_slice};
 pub use text_input::{
     longest_common_prefix, render_text_field, render_text_field_with_ghost,
